refactor: extract native recursion runtime
