+++ docs/deletion-candidates.md
@@ -7,7 +7,7 @@ later removal change.
 | Candidate | Status | Replacement | Required evidence | Fallback | Owner decision |
 |---|---|---|---|---|---|
 | Duplicated runtime policy inside `rlm_query` | retained during convergence | canonical shared runtime engine plus thin CLI adapter | contract parity, CLI async/stdin tests, consumer-pack tests, real-model eval | incumbent CLI implementation | pending |
-| Duplicated runtime policy inside `extensions/ypi/native-tool.ts` | retained during convergence | canonical shared runtime engine plus thin Pi adapter | contract parity, native harness, cancellation/output tests, real-model eval | incumbent native implementation | pending |
+| Preserved incumbent native engine in `extensions/ypi/legacy-native-tool.ts` | retained fallback during convergence | `extensions/ypi/runtime-core.ts` plus thin `extensions/ypi/native-tool.ts` Pi adapter | contract parity, native harness, cancellation/output tests, package tests, real-model eval | `YPI_LEGACY_IMPL=1` | pending |
 
 A status may move from `retained during convergence` to `marked for deletion`
 only when every required evidence item is linked in this table. Marking does not
